refactor code, update db.js file
db.js now contains three methods: getBinId, savePayload, and saveRequest.
the requestsRouter now handles logic for intaking a post request, it parses the bin path from the req object and verifies it is a valid bin_id.
Once it verifies that, it saves the payload to the mongoDb and saves the request information (including mongoid) into the psql requests table